requirements.txt
# Runtime dependencies
pyyaml==6.0.2
python-dotenv==1.0.1

# Linting and formatting
ruff==0.4.4
black==24.4.2

# Testing
pytest==8.2.0
pytest-mock==3.14.0
